Move Action from azulclient.py to queues.py
